Unify the dx8wrapper.h
